requirements.txt
streamlit
moviepy==1.0.3
numpy
Pillow
opencv-python-headless